requirements.txt
Django==4.2.5
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
flake8==6.1.0
psycopg2-binary==2.9.8
jsonschema==4.19.1
memory-profiler==0.61.0
